AsyncGenericeRepoChanges

+++ BusinesEntites/B_ForgotPassword.cs
@@ -3,7 +3,7 @@ namespace BusinesEntites
 {
     public class B_ForgotPassword
     {
-        public int MobileNo{get;set;}
+        public long MobileNo{get;set;}
         public string EmailId{get;set;}
     }
 }
